Previously, the data field would reset, now, by default the data collected thus far will be retained so that when we go back through the workflow the previously entered items will be populated.
This may cause problems in some cases where we entered a cardinality for a Sequential MI and the second time through we enter a smaller loop cardinality - we still have the old cardinality of answers in the data structure and we will need to decide which to discard. This may be done arbitrarily (i.e. last gets discarded)

In addition, we may encounter workflows that will require us to look to the finishing data for cluse about which things we need to do or not.

example:

we have a branching gateway that fills out forms one way or skips them the other way - we go through the workflow, answer all the optional field items and then we reset the token and have a situation where we skip the answers - if we do not reset the data, the previous answers are still in the target.data dictionary.