Make method that descends down in to tree, finds the nearest parallel gateway and fixes it up
I will have to do some testing to find out if this will
solve the problem I am actually trying to solve for.

(i.e. you have a matrix of tasks inside of a parallel gateway, you have done all of them and you choose [1,1] to reset the token to. )

This will likely break things if you have done row 1 and you are about to complete row 2 but you go back and reset the [2,1] to be where the token should be.