Made performance enhancement
Each time we called get_task, it was iterating through the entire tree again, but in this case we were doing it for every task in the entire tree.

This proved to be expensive, so instead, we got a copy of the task list and look through this rather than re-traversing the tree in the iterator.

If we do not pass in the tasktree, this should be transparent.