This should probably be a stash - this works well when you are not doing a save-restore, but it fails miserably when you do -
The problem is that prior to loading - it re-predicts the workflow that it already has, but without the task tree in place. Essentially, it would re-expand the task tree a second time for each time you did a save restore.

This is due to the way I was using the task to store information about the state of the task-spec tree - on save/restore, this information wasn't available on the predict pass that happens right before restoring the actual tasks.

Basically, I will need to take a whole new tack on this problem, It may be that I can add some parameters to the task spec to aid in this sort of thing - I have other problems with PMI that could use a similar treatment.