Manged to get a test case working where we are able to replace an external sub-process with a new and different subprocess by reset_token to the task just before.
I suspect that this will break if we reset to 2 or more tasks ahead of the call activity - simply because I am specifically looking for a CallActivity and snipping the tree on reset-token.

In normal operations, the task-spec tree and the task tree get 'out of sync' because the task tree reflects the actual flow including the sub-process, and the task-spec tree reflects only what is in the current workflow spec.