Add the ability to run stand alone scripts in pure python.
FEEL expression expansion is still in place - this may need to be removed if we encounter a conflict.

There was one test in cr-connect-workflow that was broken due to the new 'feelFilter' syntax, so I made the first argument be required to be non-spaces.  This implies that
x[item= 5] will work but x[item = 5] will not.