Add Tests for Feel (like) expression language
This allows us to change to feel by changing the scripting language in the DMN file. I added all of these tests with the idea that we would want them for testing the 'real' feel expression engine when we get to it. In that case it will be just a matter of swapping out the class for the Feel Expression engine where I detect that it is requesting FEEL as the engine and just re-running the tests