Add the ability for a BusinessRuleTask or a ScriptTask to be a multiinstance.
Fix problem with serialize/deserialze where a Box object was in the data. This happened because we were converting any dict object in the data into a Box object before we evaluate python code as a convenience to the configurators. The solution was to convert them back to dictionaries right after the python expression engine ran.

Lastly, there was a bug for MI that was looking for a class of sometaskname_999 because these some nasty amalgam of a mi class and the underlying base class - typically camunda will give a class a name of Task_xxxxx where xxxx is an alphanumeric - but in this case it was Task_99xxx so it was finding that _99 part and striping the underscore. I may want to make this even less likely to clash See ticket # 254 about this

This fixes:250