Fix #425 Gives a method for correcting for timezone on a timer event, but it is not automatic.
Adds a test for some of the call activity items.

You can now:

parse a date and include timezone and it should do the right thing on the server with utc timezone

dateparser.parse('2021-09-01 10:00 EDT')

and it also fixes another problem we've had with using

now() + timedelta(minutes=10)

instead you can use dateparser.parse('in 10 minutes') - and it should do the right thing.

I have not added a test for this because it doesn't parse fractions of seconds correctly, instead just yielding seconds.